Move base codestarts tests to IT to fix snapshot
