+++ integration-tests/devtools/src/test/java/io/quarkus/devtools/codestarts/quarkus/PicocliCodestartTest.java
